+++ test/validator/samples/a11y-no-noninteractive-tabindex/input.svelte
@@ -8,6 +8,7 @@
 <div role='article' tabindex='-1' />
 <article tabindex='-1' />
 <div role="tabpanel" tabindex='0' />
+<svelte:element this={Math.random() ? 'button' : 'div'} tabindex="0" />
 <!-- invalid -->
 <div tabindex='0' />
 <div role='article' tabindex='0' />