+++ test/validator/samples/a11y-scope/input.svelte
@@ -1,1 +1,6 @@
+<!-- valid -->
+<th scope />
+<svelte:element this={Math.random() ? 'th' : 'td'} scope />
+
+<!-- invalid -->
 <div scope/>